[#3] Refactor: Reduce ViewController's dependence on PhotoLibaray
